Cleanup old garbage, Get better indentation
Previously I was thinking of using temporary file to store the modified
snippet and to run that file. But I decided to change the behavior, to
open new untitled document (which has modified snippet as its content)
and invoke `code-runner.run` from there. Because of that there are many
unused codes (such as parsing file's directory / extension) in the
previous commit, so removeed them.

In addition, supports indentation of modified snippet. If `$snippet` is
naively replaced by the snippet, only the first line of the snippet will
be indented. This behavior will be a problem when using indent-based
languages (like python). Also the code which has broken indentation
looks bit ugly. So the beginning of each line of selected snippet I
append the same amount of ' ' (space) with the first line. tabs are not
supported.